Data Analyst internship project
I am thrilled to share with you all that i have successfully completed my internship at TechnoHacks Eductech as a Data Analyst role. During this internship I have worked on  different datasets to work on it and and use the data to retrieve informational insights from them to help the organisation. This was my one month internship project and i have learned alot of things and enhanced my skills in data analyst career.